+++ repository/todo_mariadb.go
@@ -14,8 +14,8 @@ func NewTodoMariaDBRepository(db *sql.DB) TodoRepository {
 	return &TodoMariaDB{db: db}
 }
 
-func (t *TodoMariaDB) GetAllTodo() ([]models.Todo, error) {
-	rows, err := t.db.Query("SELECT * FROM todos")
+func (t *TodoMariaDB) GetAllTodo(userId int64) ([]models.Todo, error) {
+	rows, err := t.db.Query("SELECT * FROM todos where user_id = ?", userId)
 	if err != nil {
 		return nil, err
 	}
@@ -23,13 +23,14 @@ func (t *TodoMariaDB) GetAllTodo() ([]models.Todo, error) {
 	var todos []models.Todo
 	for rows.Next() {
 		var id int64
+		var userId int64
 		var title string
 		var description string
-		err = rows.Scan(&id, &title, &description)
+		err = rows.Scan(&id, &userId, &title, &description)
 		if err != nil {
 			return nil, err
 		}
-		todos = append(todos, models.Todo{Id: id, Title: title, Description: description})
+		todos = append(todos, models.Todo{Id: id, UserId: userId, Title: title, Description: description})
 	}
 
 	return todos, nil
@@ -37,9 +38,10 @@ func (t *TodoMariaDB) GetAllTodo() ([]models.Todo, error) {
 
 func (t *TodoMariaDB) GetTodo(todo models.Todo) (models.Todo, error) {
 	var id int64
+	var userId int64
 	var title string
 	var description string
-	err := t.db.QueryRow("SELECT * FROM todos WHERE id = ?", todo.Id).Scan(&id, &title, &description)
+	err := t.db.QueryRow("SELECT * FROM todos WHERE id = ? AND user_id = ?", todo.Id, todo.UserId).Scan(&id, &userId, &title, &description)
 	if err != nil {
 		return models.Todo{}, err
 	}
@@ -48,7 +50,7 @@ func (t *TodoMariaDB) GetTodo(todo models.Todo) (models.Todo, error) {
 }
 
 func (t *TodoMariaDB) AddTodo(todo models.Todo) (models.Todo, error) {
-	result, err := t.db.Exec("INSERT INTO todos (title, description) VALUES (?, ?)", todo.Title, todo.Description)
+	result, err := t.db.Exec("INSERT INTO todos (title, user_id, description) VALUES (?, ?, ?)", todo.Title, todo.UserId, todo.Description)
 	if err != nil {
 		return models.Todo{}, err
 	}
@@ -73,7 +75,7 @@ func (t *TodoMariaDB) AddTodo(todo models.Todo) (models.Todo, error) {
 }
 
 func (t *TodoMariaDB) DeleteTodo(todo models.Todo) (models.Todo, error) {
-	result, err := t.db.Exec("DELETE FROM todos WHERE id = ?", todo.Id)
+	result, err := t.db.Exec("DELETE FROM todos WHERE id = ? AND user_id = ?", todo.Id, todo.UserId)
 	if err != nil {
 		return models.Todo{}, err
 	}
